backend/video_token_project/requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.26.0
django-jazzmin>=2.6
